refactor: decomposed a dispatcher component for a better readability
